rvv: fp16: support conversion instrucitons

+++ riscv/insns/vfcvt_f_x_v.h
@@ -1,7 +1,7 @@
 // vfcvt.f.x.v vd, vd2, vm
 VI_VFP_VF_LOOP
 ({
-  auto vs2_i = P.VU.elt<int32_t>(rs2_num, i);
+  auto vs2_i = P.VU.elt<int16_t>(rs2_num, i);
   vd = i32_to_f16(vs2_i);
 },
 {
