+++ riscv/insns/vfcvt_f_xu_v.h
@@ -1,7 +1,7 @@
 // vfcvt.f.xu.v vd, vd2, vm
 VI_VFP_VF_LOOP
 ({
-  auto vs2_u = P.VU.elt<uint32_t>(rs2_num, i);
+  auto vs2_u = P.VU.elt<uint16_t>(rs2_num, i);
   vd = ui32_to_f16(vs2_u);
 },
 {
